refactor(triage): isolate Claude capability probe
