+++ src/commands/triage.test.ts
@@ -1031,33 +1031,6 @@ describe("triageCommand", () => {
     }
   });
 
-  it("prints a manual handoff instead of launching Claude without safe-mode support", async () => {
-    mocks.resolveExecutablePath.mockImplementation((binary: string) =>
-      binary === "claude" ? "/usr/local/bin/claude" : undefined,
-    );
-    mocks.runUtf8CommandWithTimeout.mockResolvedValue({
-      stdout: "Usage: claude [options]",
-      stderr: "",
-      code: 0,
-      termination: "exit",
-    });
-    const runtime = createTriageRuntime();
-
-    await withTriageTerminal(true, async () => {
-      await expect(triageCommand(runtime, { noExport: true })).rejects.toMatchObject({ code: 1 });
-    });
-
-    expect(mocks.spawn).not.toHaveBeenCalled();
-    expect(mocks.runUtf8CommandWithTimeout).toHaveBeenCalledWith(
-      ["/usr/local/bin/claude", "--help"],
-      expect.objectContaining({ timeoutMs: 10_000, killProcessTree: true }),
-    );
-    expect(runtime.error).toHaveBeenCalledWith(
-      expect.stringContaining("Claude Code 2.1.169 or newer is required"),
-    );
-    expect(runtime.log).toHaveBeenCalledWith(expect.stringContaining("Run without safe mode:"));
-  });
-
   it("reports a failed launch without trying another installed agent", async () => {
     mocks.resolveExecutablePath.mockImplementation((binary: string) => `/usr/local/bin/${binary}`);
     mocks.spawn.mockImplementation(() => {
